requirements.txt
requests
textblob
python-dotenv
python-telegram-bot
vaderSentiment
flask
